// Student/backend/src/mock-test/dto/create-staff-mock-test.dto.ts
export class CreateStaffMockTestDto {
  studentId?: string;
  studentName?: string;
  skill?: string;
  day?: number;
  month?: number;
  year?: number;
  examTime?: string;
  examTeacher?: string;
  examLink?: string;
  note?: string;
  status?: string;
  guestPhone?: string;
  leadId?: string;
  source?: string;
  entranceBookingId?: string;
  finalTestId?: string;
}